requirements.txt
# OpenHands RL Agent 依赖
# 基于 openhands-magic (sft) 的环境配置

# OpenHands SDK 核心（必须使用 1.8.2 版本）
openhands-sdk==1.8.2
openhands-tools==1.8.2
openhands-agent-server==1.8.2
openhands-workspace==1.8.2

# LiteLLM（必须使用 1.80.16 版本，避免 Responses API 兼容性问题）
litellm==1.80.16

# OpenAI 客户端
openai==2.8.0

# HTTP 客户端
httpx==0.28.1

# 数据处理
pydantic>=2.12.5
python-dotenv>=1.2.1

# RL 训练依赖（openhands 环境需要）
torch>=2.10.0
transformers>=5.1.0
trl>=0.28.0
accelerate>=1.12.0
datasets>=4.5.0
peft>=0.18.1

# 其他依赖
tenacity>=9.1.2
pyyaml>=6.0.3
requests>=2.31.0
